split course service
