+++ src/main/java/de/tum/cit/aet/artemis/core/web/admin/AdminCourseResource.java
@@ -41,8 +41,11 @@ import de.tum.cit.aet.artemis.core.exception.BadRequestAlertException;
 import de.tum.cit.aet.artemis.core.repository.CourseRepository;
 import de.tum.cit.aet.artemis.core.repository.UserRepository;
 import de.tum.cit.aet.artemis.core.security.annotations.EnforceAdmin;
-import de.tum.cit.aet.artemis.core.service.CourseService;
 import de.tum.cit.aet.artemis.core.service.FileService;
+import de.tum.cit.aet.artemis.core.service.course.CourseAccessService;
+import de.tum.cit.aet.artemis.core.service.course.CourseAdminService;
+import de.tum.cit.aet.artemis.core.service.course.CourseDeletionService;
+import de.tum.cit.aet.artemis.core.service.course.CourseService;
 import de.tum.cit.aet.artemis.core.util.FilePathConverter;
 import de.tum.cit.aet.artemis.core.util.FileUtil;
 import de.tum.cit.aet.artemis.core.util.HeaderUtil;
@@ -61,12 +64,16 @@ public class AdminCourseResource {
 
     private static final int MAX_TITLE_LENGTH = 255;
 
+    private final CourseAccessService courseAccessService;
+
     @Value("${jhipster.clientApp.name}")
     private String applicationName;
 
     private final UserRepository userRepository;
 
     private final CourseService courseService;
+
+    private final CourseAdminService courseAdminService;
 
     private final ChannelService channelService;
 
@@ -78,15 +85,21 @@ public class AdminCourseResource {
 
     private final Optional<LtiApi> ltiApi;
 
-    public AdminCourseResource(UserRepository userRepository, CourseService courseService, CourseRepository courseRepository, AuditEventRepository auditEventRepository,
-            FileService fileService, Optional<LtiApi> ltiApi, ChannelService channelService) {
+    private final CourseDeletionService courseDeletionService;
+
+    public AdminCourseResource(UserRepository userRepository, CourseService courseService, CourseAdminService courseAdminService, CourseRepository courseRepository,
+            AuditEventRepository auditEventRepository, FileService fileService, Optional<LtiApi> ltiApi, ChannelService channelService, CourseDeletionService courseDeletionService,
+            CourseAccessService courseAccessService) {
         this.courseService = courseService;
+        this.courseAdminService = courseAdminService;
         this.courseRepository = courseRepository;
         this.auditEventRepository = auditEventRepository;
         this.userRepository = userRepository;
         this.fileService = fileService;
         this.ltiApi = ltiApi;
         this.channelService = channelService;
+        this.courseDeletionService = courseDeletionService;
+        this.courseAccessService = courseAccessService;
     }
 
     /**
@@ -147,7 +160,7 @@ public class AdminCourseResource {
             ltiApi.get().createOnlineCourseConfiguration(course);
         }
 
-        courseService.setDefaultGroupsIfNotSet(course);
+        courseAccessService.setDefaultGroupsIfNotSet(course);
 
         Course createdCourse = courseRepository.save(course);
 
@@ -179,7 +192,7 @@ public class AdminCourseResource {
         auditEventRepository.add(auditEvent);
         log.info("User {} has requested to delete the course {}", user.getLogin(), course.getTitle());
 
-        courseService.delete(course);
+        courseDeletionService.delete(course);
         if (course.getCourseIcon() != null) {
             fileService.schedulePathForDeletion(FilePathConverter.fileSystemPathForExternalUri(URI.create(course.getCourseIcon()), FilePathType.COURSE_ICON), 0);
         }
@@ -195,7 +208,7 @@ public class AdminCourseResource {
     @GetMapping("courses/{courseId}/deletion-summary")
     public ResponseEntity<CourseDeletionSummaryDTO> getDeletionSummary(@PathVariable long courseId) {
         log.debug("REST request to get deletion summary course: {}", courseId);
-        return ResponseEntity.ok().body(courseService.getDeletionSummary(courseId));
+        return ResponseEntity.ok().body(courseAdminService.getDeletionSummary(courseId));
     }
 
     /**
