+++ src/test/java/de/tum/cit/aet/artemis/core/service/CourseStatsServiceTest.java
@@ -1,0 +1,140 @@
+package de.tum.cit.aet.artemis.core.service;
+
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.stream.IntStream;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.springframework.beans.factory.annotation.Autowired;
+
+import de.tum.cit.aet.artemis.core.domain.Language;
+import de.tum.cit.aet.artemis.core.security.SecurityUtils;
+import de.tum.cit.aet.artemis.core.service.course.CourseStatsService;
+import de.tum.cit.aet.artemis.core.user.util.UserUtilService;
+import de.tum.cit.aet.artemis.exercise.domain.participation.StudentParticipation;
+import de.tum.cit.aet.artemis.exercise.repository.ExerciseRepository;
+import de.tum.cit.aet.artemis.shared.base.AbstractSpringIntegrationIndependentTest;
+import de.tum.cit.aet.artemis.text.domain.TextSubmission;
+import de.tum.cit.aet.artemis.text.util.TextExerciseFactory;
+
+public class CourseStatsServiceTest extends AbstractSpringIntegrationIndependentTest {
+
+    private static final String TEST_PREFIX = "coursestatsservice";
+
+    @Autowired
+    private CourseStatsService courseStatsService;
+
+    @Autowired
+    private UserUtilService userUtilService;
+
+    @Autowired
+    private ExerciseRepository exerciseRepository;
+
+    @BeforeEach
+    void initTestCase() {
+        userUtilService.addUsers(TEST_PREFIX, 2, 0, 0, 1);
+    }
+
+    static IntStream weekRangeProvider() {
+        // test all weeks (including the year change from 52 or 53 to 0)
+        return IntStream.range(0, 60);
+    }
+
+    @ParameterizedTest(name = "{displayName} [{index}] {argumentsWithNames}")
+    @MethodSource("weekRangeProvider")
+    void testGetActiveStudents(long weeks) {
+        ZonedDateTime date = ZonedDateTime.now().minusWeeks(weeks);
+        SecurityUtils.setAuthorizationObject();
+        var course = courseUtilService.addEmptyCourse();
+        var exercise = TextExerciseFactory.generateTextExercise(date, date, date, course);
+        course.addExercises(exercise);
+        exercise = exerciseRepository.save(exercise);
+
+        var student1 = userUtilService.getUserByLogin(TEST_PREFIX + "student1");
+        var participation1 = new StudentParticipation();
+        participation1.setParticipant(student1);
+        participation1.exercise(exercise);
+        var student2 = userUtilService.getUserByLogin(TEST_PREFIX + "student2");
+        var participation2 = new StudentParticipation();
+        participation2.setParticipant(student2);
+        participation2.exercise(exercise);
+        var participation3 = new StudentParticipation();
+        participation3.setParticipant(student2);
+        participation3.exercise(exercise);
+        var participation4 = new StudentParticipation();
+        participation4.setParticipant(student2);
+        participation4.exercise(exercise);
+        studentParticipationRepo.saveAll(Arrays.asList(participation1, participation2, participation3, participation4));
+
+        var submission1 = new TextSubmission();
+        submission1.text("text of text submission1");
+        submission1.setLanguage(Language.ENGLISH);
+        submission1.setSubmitted(true);
+        submission1.setParticipation(participation1);
+        submission1.setSubmissionDate(date);
+
+        var submission2 = new TextSubmission();
+        submission2.text("text of text submission2");
+        submission2.setLanguage(Language.ENGLISH);
+        submission2.setSubmitted(true);
+        submission2.setParticipation(participation2);
+        submission2.setSubmissionDate(date);
+
+        var submission3 = new TextSubmission();
+        submission3.text("text of text submission3");
+        submission3.setLanguage(Language.ENGLISH);
+        submission3.setSubmitted(true);
+        submission3.setSubmissionDate(date.minusDays(14));
+        submission3.setParticipation(participation3);
+
+        var submission4 = new TextSubmission();
+        submission4.text("text of text submission4");
+        submission4.setLanguage(Language.ENGLISH);
+        submission4.setSubmitted(true);
+        submission4.setSubmissionDate(date.minusDays(7));
+        submission4.setParticipation(participation4);
+
+        submissionRepository.saveAll(Arrays.asList(submission1, submission2, submission3, submission4));
+
+        var exerciseList = new HashSet<Long>();
+        exerciseList.add(exercise.getId());
+        var activeStudents = courseStatsService.getActiveStudents(exerciseList, 0, 4, date);
+        assertThat(activeStudents).hasSize(4).containsExactly(0, 1, 1, 2);
+    }
+
+    @Test
+    void testGetActiveStudents_UTCConversion() {
+        ZonedDateTime date = ZonedDateTime.of(2022, 1, 2, 0, 0, 0, 0, ZonedDateTime.now().getZone());
+        SecurityUtils.setAuthorizationObject();
+        var course = courseUtilService.addEmptyCourse();
+        var exercise = TextExerciseFactory.generateTextExercise(date, date, date, course);
+        course.addExercises(exercise);
+        exercise = exerciseRepository.save(exercise);
+
+        var student1 = userUtilService.getUserByLogin(TEST_PREFIX + "student1");
+        var participation1 = new StudentParticipation();
+        participation1.setParticipant(student1);
+        participation1.exercise(exercise);
+
+        studentParticipationRepo.save(participation1);
+
+        var submission1 = new TextSubmission();
+        submission1.text("text of text submission1");
+        submission1.setLanguage(Language.ENGLISH);
+        submission1.setSubmitted(true);
+        submission1.setParticipation(participation1);
+        submission1.setSubmissionDate(date.plusDays(1).plusMinutes(59).plusSeconds(59).plusNanos(59));
+
+        submissionRepository.save(submission1);
+
+        var exerciseList = new HashSet<Long>();
+        exerciseList.add(exercise.getId());
+        var activeStudents = courseStatsService.getActiveStudents(exerciseList, 0, 4, ZonedDateTime.of(2022, 1, 25, 0, 0, 0, 0, ZoneId.systemDefault()));
+        assertThat(activeStudents).hasSize(4).containsExactly(1, 0, 0, 0);
+    }
+}
